Added an rqt plugin for running the demo.

+++ branches/experimental/stacks/bwi/camera_person_detector/include/detector/DetectorTypeDefs.h
@@ -1,1 +1,1 @@
-typedef uint64_t GUID;
+typedef uint32_t GUID;
